requirements.txt
# Core dependencies for Snake Game
pygame>=2.5.0
numpy>=1.24.0
gradio>=4.0.0
Pillow>=10.0.0

# Additional dependencies for game functionality
typing-extensions>=4.0.0

# Development and debugging tools
ipython>=8.0.0
pytest>=7.0.0

# Optional for future enhancements
requests>=2.28.0